contentflow-ai/src/tools/crawl4ai_tools.py
"""Tools for web content extraction using crawl4ai.

This module provides tools for extracting content from websites using the
crawl4ai library. These tools can be used by the extraction agents to retrieve
content from various web sources.
"""

from typing import Dict, List, Optional, Any
from google.adk.tools import Tool


def extract_web_content(url: str, selectors: Optional[Dict[str, str]] = None) -> Dict[str, Any]:
    """
    Extract content from a web page using crawl4ai.

    Args:
        url: The URL of the web page to extract content from.
        selectors: Optional CSS selectors to extract specific content.
            Format: {'title': 'h1', 'content': 'article', ...}

    Returns:
        A dictionary containing the extracted content with metadata.
    """
    # This is a placeholder implementation
    # In a real implementation, we would use crawl4ai to extract the content
    return {
        "url": url,
        "title": "Sample Title",
        "content": "Sample content extracted from the web page.",
        "metadata": {
            "source": url,
            "extraction_time": "2025-04-15T01:00:00Z",
            "word_count": 100,
        }
    }


def extract_article_content(url: str) -> Dict[str, Any]:
    """
    Extract article content from a web page using crawl4ai.
    
    This function is specialized for article extraction, focusing on
    title, author, publication date, and main content.

    Args:
        url: The URL of the article to extract content from.

    Returns:
        A dictionary containing the extracted article content with metadata.
    """
    # This is a placeholder implementation
    # In a real implementation, we would use crawl4ai to extract the article
    return {
        "url": url,
        "title": "Sample Article Title",
        "author": "John Doe",
        "published_date": "2025-04-10",
        "content": "Sample article content extracted from the web page.",
        "images": ["https://example.com/image1.jpg"],
        "metadata": {
            "source": url,
            "extraction_time": "2025-04-15T01:00:00Z",
            "word_count": 500,
            "reading_time": "3 minutes",
        }
    }


# Create ADK tools for web content extraction
web_content_extraction_tool = Tool(
    name="extract_web_content",
    description="Extracts content from a web page using crawl4ai",
    function=extract_web_content
)

article_extraction_tool = Tool(
    name="extract_article_content",
    description="Extracts article content from a web page using crawl4ai",
    function=extract_article_content
)
